# Conflicts:
#	Data/JamesDAL.Client/schema.graphql
#	Data/JamesDAL.Server/GraphQL/Mutations/GeneralMutation.cs
#	Data/JamesDAL.Server/GraphQL/Queries/General.cs
#	JamesWebUI/Client/Components/AgencyComponents/AgencyAgentsGrid.razor
#	JamesWebUI/Client/Pages/AgencyProfile.razor
#	JamesWebUI/Client/Shared/MainLayout.razor

+++ Shared/Model/Ratio.cs
@@ -24,169 +24,169 @@ public partial class Ratio
 
     public string? StatementQuality { get; set; }
 
-    public int? WorkingCapitalStated { get; set; }
+    public long? WorkingCapitalStated { get; set; }
 
-    public int? TotalAssetsStated { get; set; }
+    public long? TotalAssetsStated { get; set; }
 
-    public int? RetainedEarningsStated { get; set; }
+    public long? RetainedEarningsStated { get; set; }
 
-    public int? EarningsBeforeTaxesStated { get; set; }
+    public long? EarningsBeforeTaxesStated { get; set; }
 
-    public int? NetWorthStated { get; set; }
+    public long? NetWorthStated { get; set; }
 
-    public int? TotalLiabilitiesStated { get; set; }
+    public long? TotalLiabilitiesStated { get; set; }
 
-    public int? TotalEquityStated { get; set; }
+    public long? TotalEquityStated { get; set; }
 
-    public int? TotalRevenueStated { get; set; }
+    public long? TotalRevenueStated { get; set; }
 
-    public int? TotalCostOfGoodsSoldStated { get; set; }
+    public long? TotalCostOfGoodsSoldStated { get; set; }
 
-    public int? CurrentAssetsStated { get; set; }
+    public long? CurrentAssetsStated { get; set; }
 
-    public int? CurrentLiabilitiesStated { get; set; }
+    public long? CurrentLiabilitiesStated { get; set; }
 
-    public int? NetIncomeStated { get; set; }
+    public long? NetIncomeStated { get; set; }
 
-    public int? InventoryStated { get; set; }
+    public long? InventoryStated { get; set; }
 
-    public int? FixedAssetsStated { get; set; }
+    public long? FixedAssetsStated { get; set; }
 
-    public int? AccountsReceivableStated { get; set; }
+    public long? AccountsReceivableStated { get; set; }
 
-    public int? AccountsPayableStated { get; set; }
+    public long? AccountsPayableStated { get; set; }
 
-    public int? SubordinatedNotesStated { get; set; }
+    public long? SubordinatedNotesStated { get; set; }
 
-    public int? RetainageReceivedStated { get; set; }
+    public long? RetainageReceivedStated { get; set; }
 
-    public int? RetainagePayableStated { get; set; }
+    public long? RetainagePayableStated { get; set; }
 
-    public int? CashPlusEarningsGreaterThanBillingsStated { get; set; }
+    public long? CashPlusEarningsGreaterThanBillingsStated { get; set; }
 
-    public int? BillingsGreaterThanCostsPlusEarningsStated { get; set; }
+    public long? BillingsGreaterThanCostsPlusEarningsStated { get; set; }
 
-    public int? PrepaidExpensesStated { get; set; }
+    public long? PrepaidExpensesStated { get; set; }
 
-    public int? CashInventoryStated { get; set; }
+    public long? CashInventoryStated { get; set; }
 
-    public int? OtherCurrentAssetsStated { get; set; }
+    public long? OtherCurrentAssetsStated { get; set; }
 
-    public int? BankDebtStated { get; set; }
+    public long? BankDebtStated { get; set; }
 
-    public int? GaexpensesStated { get; set; }
+    public long? GaexpensesStated { get; set; }
 
-    public int? OperatingProfitStated { get; set; }
+    public long? OperatingProfitStated { get; set; }
 
-    public int? IncomeAfterTaxesStated { get; set; }
+    public long? IncomeAfterTaxesStated { get; set; }
 
-    public int? CashStated { get; set; }
+    public long? CashStated { get; set; }
 
-    public int? NotesStated { get; set; }
+    public long? NotesStated { get; set; }
 
-    public int? QscoreStated { get; set; }
+    public long? QscoreStated { get; set; }
 
-    public int? QprofitStated { get; set; }
+    public long? QprofitStated { get; set; }
 
-    public int? QnetworthStated { get; set; }
+    public long? QnetworthStated { get; set; }
 
-    public int? QleverageStated { get; set; }
+    public decimal? QleverageStated { get; set; }
 
-    public int? QworkingCapitalStated { get; set; }
+    public long? QworkingCapitalStated { get; set; }
 
-    public int? QcratioStated { get; set; }
+    public decimal? QcratioStated { get; set; }
 
-    public int? WorkingCapitalAllowed { get; set; }
+    public long? WorkingCapitalAllowed { get; set; }
 
-    public int? TotalAssetsAllowed { get; set; }
+    public long? TotalAssetsAllowed { get; set; }
 
-    public int? RetainedEarningsAllowed { get; set; }
+    public long? RetainedEarningsAllowed { get; set; }
 
-    public int? EarningsBeforeTaxesAllowed { get; set; }
+    public long? EarningsBeforeTaxesAllowed { get; set; }
 
-    public int? NetWorthAllowed { get; set; }
+    public long? NetWorthAllowed { get; set; }
 
-    public int? TotalLiabilitiesAllowed { get; set; }
+    public long? TotalLiabilitiesAllowed { get; set; }
 
-    public int? TotalEquityAllowed { get; set; }
+    public long? TotalEquityAllowed { get; set; }
 
-    public int? TotalRevenueAllowed { get; set; }
+    public long? TotalRevenueAllowed { get; set; }
 
-    public int? TotalCostOfGoodsSoldAllowed { get; set; }
+    public long? TotalCostOfGoodsSoldAllowed { get; set; }
 
-    public int? CurrentAssetsAllowed { get; set; }
+    public long? CurrentAssetsAllowed { get; set; }
 
-    public int? CurrentLiabilitiesAllowed { get; set; }
+    public long? CurrentLiabilitiesAllowed { get; set; }
 
-    public int? NetIncomeAllowed { get; set; }
+    public long? NetIncomeAllowed { get; set; }
 
-    public int? InventoryAllowed { get; set; }
+    public long? InventoryAllowed { get; set; }
 
-    public int? FixedAssetsAllowed { get; set; }
+    public long? FixedAssetsAllowed { get; set; }
 
-    public int? AccountsReceivableAllowed { get; set; }
+    public long? AccountsReceivableAllowed { get; set; }
 
-    public int? AccountsPayableAllowed { get; set; }
+    public long? AccountsPayableAllowed { get; set; }
 
-    public int? SubordinatedNotesAllowed { get; set; }
+    public long? SubordinatedNotesAllowed { get; set; }
 
-    public int? RetainageReceivedAllowed { get; set; }
+    public long? RetainageReceivedAllowed { get; set; }
 
-    public int? RetainagePayableAllowed { get; set; }
+    public long? RetainagePayableAllowed { get; set; }
 
-    public int? CashPlusEarningsGreaterThanBillingsAllowed { get; set; }
+    public long? CashPlusEarningsGreaterThanBillingsAllowed { get; set; }
 
-    public int? BillingsGreaterThanCostsPlusEarningsAllowed { get; set; }
+    public long? BillingsGreaterThanCostsPlusEarningsAllowed { get; set; }
 
-    public int? PrepaidExpensesAllowed { get; set; }
+    public long? PrepaidExpensesAllowed { get; set; }
 
-    public int? CashInventoryAllowed { get; set; }
+    public long? CashInventoryAllowed { get; set; }
 
-    public int? OtherCurrentAssetsAllowed { get; set; }
+    public long? OtherCurrentAssetsAllowed { get; set; }
 
-    public int? BankDebtAllowed { get; set; }
+    public long? BankDebtAllowed { get; set; }
 
-    public int? GaexpensesAllowed { get; set; }
+    public long? GaexpensesAllowed { get; set; }
 
-    public int? OperatingProfitAllowed { get; set; }
+    public long? OperatingProfitAllowed { get; set; }
 
-    public int? IncomeAfterTaxesAllowed { get; set; }
+    public long? IncomeAfterTaxesAllowed { get; set; }
 
-    public int? CashAllowed { get; set; }
+    public long? CashAllowed { get; set; }
 
-    public int? NotesAllowed { get; set; }
+    public long? NotesAllowed { get; set; }
 
-    public int? QscoreAllowed { get; set; }
+    public long? QscoreAllowed { get; set; }
 
-    public int? QprofitAllowed { get; set; }
+    public long? QprofitAllowed { get; set; }
 
-    public int? QnetworthAllowed { get; set; }
+    public long? QnetworthAllowed { get; set; }
 
-    public int? QleverageAllowed { get; set; }
+    public decimal? QleverageAllowed { get; set; }
 
-    public int? QworkingCapitalAllowed { get; set; }
+    public long? QworkingCapitalAllowed { get; set; }
 
-    public int? QcratioAllowed { get; set; }
+    public decimal? QcratioAllowed { get; set; }
 
     public string? Scaling { get; set; }
 
-    public int? SingleLoa { get; set; }
+    public long? SingleLoa { get; set; }
 
-    public int? AggregateLoa { get; set; }
+    public long? AggregateLoa { get; set; }
 
-    public int? LargestJob { get; set; }
+    public long? LargestJob { get; set; }
 
-    public int? LargestBacklog { get; set; }
+    public long? LargestBacklog { get; set; }
 
-    public int? CashFromOperations { get; set; }
+    public long? CashFromOperations { get; set; }
 
-    public int? CashFromInvestments { get; set; }
+    public long? CashFromInvestments { get; set; }
 
-    public int? CashFromFinancing { get; set; }
+    public long? CashFromFinancing { get; set; }
 
     public string? Comments { get; set; }
 
-    public int? NetChangeCash { get; set; }
+    public long? NetChangeCash { get; set; }
 
     public Guid? ImagingId { get; set; }
